Fix orphan zap map bug in LDMS re-lookup
When an application re-lookup an LDMS set with LDMS_LOOKUP_SET_INFO
flag, the returning rendezvous zap map is a new zap map, which is
expected to be unmapped if not used. The `handle_rendezvous_lookup()`
routine left the new rendezvous map orphan (no assoicated LDMS RBD),
which will hold a reference to zap xprt, resulting in lingering zap
xprt. The new zap rendezvous map is now unmapped if not used.