ldms_xprt rbd collection using red-black tree
An rbd lookup in an LDMS xprt using LDMS set occurs when:

- issuing a remote lookup,
- handle rendezvous lookup,
- handle rendezvous push.

When we have a lot of sets for a transport (e.g. agg2 -- agg1
connection), the linear lookup can take a lot of time -- especially
initally when ldmsd lookup all sets.

This patch removes `rbd_list` (linked list) and replace it with
`rbd_rbt` (red-black tree) to improve lookup performance.