Fix prdcr_set ref put in hint tree update
Hint tree update REMOVE op removes the prdcr_set from the hint list.
The REMOVE can be invoked twice for a given prdcr_set in the case that
ldmsd receives DIR_DEL while having outsanding SET_INFO lookup. The
DIR_DEL drives the first REMOVE, and the update completion drives the
second REMOVE. The existing guarding logic only guards the
LIST_REMOVE, while the prdcr_set reference is put regardless. This
exta ref put crashes ldmsd.

In this patch, the prdcr_set reference is put only when the set is
removed from the hint list. (This is the reference we took when
inserting the set into the list).