Fix cfgobj ref in delete path (double free)
When a config object reference reached 0, a delete routine is invoked.
The delete routine cleanup the object first (e.g. free `name`
attribute, etc.), then a global object tree lock is taken and the
object is removed from the config object tree. Hence, the object is
accessible in a cleaned up state. This causes a double-free issue in
the following scenario:

- thread1 locks the tree mutex and iterating through config objects.
  Let's name our config object in question as `cfgobj`. `cfgobj` has a
  reference of 1 (for being in the tree).
  - the iterator has not reached `cfgobj` yet.
- thread2 asynchronously deletes `cfgobj`, dropping the reference to
  0.  The delete routine is then invoked. The object is cleaned up.
  Then, it takes the tree mutex for removing the `cfgobj` from the
  tree.  Since thread1 held the mutex, this will block thread2.
- The iterator on thread1 reaches `cfgobj`, increasing the reference
  from 0 to 1 ... and later reduces the reference from 1 back to 0
  when it moves to the next object. The delete routine is invoked for
  the 2nd time, and so is the object cleanup ... a double free.
  - Additional problem includes:
    - accessing freed memory
    - if the double free was not detected first, a deadlock will occur
      when thread1 try to remove the object from the tree.

This patch addresses the issue by taking the global lock in the delete
request handling, and remove it from the tree before putting down the
tree reference (after checking the state and what not), so that it
won't be accessible by find / iteration. The object is then cleaned up
and freed when reference reaches 0. This time we're sure that no one
is referring to it.

This patch also introduce `STOPPING` state to updtr to track task
stopping and joining, as the task also refers to updtr object.
`STOPPING` ensures that updtr delete will fail (with EBUSY) as the
updtr has not completely stopped.