Fix ovis_event use-after-free bug
After an application callback has completed in the epoll event case,
ovis scheduler access the event object to check if it has a timer on,
and re-schedule the time out as appropriate. However, the event object
could aready be freed by the callback function. A use-after-free.
The scheduler should never use the event object after the callback
function. This patch addresses the issue by checking/recalculating the
timeout event before calling the callback function.