ldmsd: listening endpoint with specific auth and listen cfgobj
Instead of having only the default authentication method, this change enables
LDMSD to support an authentication method for each listening endpoint. The
'listen' cfgobj type is added to contain a listening endpoint configuration. A
listening endpoint can be configured 1) on command line, 2) in configuration
files or 3) via an ldmsd interface.

The listening endpoints specified on command line will always use the default
authentication method. The listening endpoints specified in configuration files
or via an ldmsd interface could have its own authentication method.

Example

listen xprt=<xprt> [port=port host=host auth=auth <auth opts>]

If the port is not given, the default port is used.
If the auth is not given, the default authentication method is used. If the
default authentication method is not specified, 'none' is used.