Fix top-level build issue
The new libraries (ev and json) in `lib/` have not been added to the
list of `OPTION_WITH_OR_BUILD()` macro calling in `ldms/`. This resulted
in `ld` not finding `libev` and `libjson_util` libraries when building
from the top of OVIS tree. This patch added ev and json into the list
and now the top-level build works.

NOTE: Also tested with distributed tar ball.