ldmsd: cleanup error messages in the config file and include paths
Error messages in any configuration files either from the cmd-line option -c or
an include request are reported in the log file (if none is given, it will be
printed to standard out). The message consists of the file path, the line number
and the acture error message.

In case of an include request from an LDMSD interface, only the message saying
that "there is an error in the included file. Please see the detail in the log
file" will be reported back to the interactive session.