Add set meta-data information to the dir reply message
The LDMS_DIR_ messages inform the client of the list
of sets published by a peer. It is also used to inform
the client of changes to this list. This patch adds
additional information about the set which will allow
clients such as ldmsd to make decisions about whether
or not to update the set without instantiating the set
in memory and allocating the resources required to do so.