ldms_xprt forwards DIR_UPD to clients, bug fixes and code cleanup
With this change, ldms_xprt sends a DIR_UPD to clients in the DIR_UPD and lookup
callback path if the set info is changed.

The lookup callback path needs to send a DIR_UPD so that if the daemon doesn't
register for DIR_F_NOTIFY its client still receives DIR_UPD in case of there are
changes in the set info.