ldmsd: refactor the set group logic to use the new dir protocol
Dir reply/update messages contain the set info data and LDMS transport forwards
the DIR_UPD if there are any changes in the set_info. The generation number in
the set info and the generation number metric are removed. The member list
update is driven by the dir_upd message instead of the comparison between the gn
in the set info and the gn metric.