Make set_id a unique verify-able number
The set_id passed in network messages was a pointer to a data
structure. There was no way to verify that the associated
data structure still existed or had been freed and reused.

This change creates a format 'set-id' that can be unambiguously
looked up to determine if the set still exists.