Improve git sha resolution
In the case that `build` directory is not in the source tree, our build
script cannot resolve git SHA as `git` invoked from the build directory
is not in a git working tree. This patch improves the script by changing
the directory to source tree (in the subshell) before calling git
commands.