Handle stuck jobs in papi_sampler sample() state machine
Jobs can get stuck in the IDLE, STOPPING, and ERROR states if
the slurmstepd or application crashes resulting in events being
dropped.

Jobs in the RUNNING state should get PAPI errors which will
transition the job to the ERROR state and prevent
the job from becoming stuck if the process goes away.