papi_sampler pause/resume syspapi
If both syspapi and papi_sampler were both active, the counters in
papi_sampler could be all 0 without any PAPI warnings or errors. This
patch add a feature in papi_sampler to automatically pause syspapi
when a PAPI-enabled job is active, and to resume syspapi when all
PAPI-enabled jobs have stopped.

syspapi can be configured to ignore the notification (auto_pause=0),
but the default is to enable the auto_pause feature (auto_pause=1).

NOTE: This patch also change the default `cumulative` option of
syspapi to 0 (do not accumulate counters) to make it consistent with
papi_sampler.