Fix dangling set on level-2 aggregator
Considering the following scenario: 4 samplers => 2 L1 agg => 1 L2 agg.
When one of the samplers has terminated, the L1 agg collecting sets from
that sampler recognized that the producer has gone and producer sets
were reset. As expected, the sets from the terminated sampler were
removed from L1 agg. However, L2 agg still have the sets around.

The cause of the problem is that the "update_ev" of the prd_set was
always reschedules even though the LDMS update completion results in an
error. This kept the prd_set reference always > 0, and prd_set stayed
forever alive but stuck in "UPDATING" state.

This patch introduces "ERROR" state to the prd_set so that the update
actor knows that the update resulted in an error and stop update event
rescheduling.