Relocate json_util library from lib to ldms
Move the json_util libary from lib to ldms and change it from
an installed library to a noinst convenience library.  Likewise
the associated header is changed to stay in-tree.

The json_test binary is changed from sbin_PROGRAMS
to check_PROGRAMS.

This is in support of issue #71, to clean up the installed
namespace footprint of ldms. It is also aligned with the desire to
eliminate recursive autoconf infrastructure.