Fix test_sampler metric increment logic
The `test_sampler` stores the latest value of the metric in `schema`
structure, which is shared among sets. Hence, when there are more than
one sets, the value of the metrics got increment by the number of sets
instead of 1. This patch modifies the increment logic to use the metric
value from the sets.