Add version string in ldmsd for binary string dump check
`ldmsd -V` shows all the version numbers that we need, but `ldms-test`
has a use case that needs to determine the version of the binary in the
environment that could not execute that binary. This patch will make
"LDMSD_VERSION #.#.#" presents in `ldmsd` and could be check with
`strings /opt/ovis/sbin/ldmsd | grep LDMSD_VERSION`.