Markdown link fix
TLDR; Make internal hyper links in Markdown files work on GitHub, GitLab
and in Doxygen-genarated HTML.

Historically, Doxygen does not understand relative document location in
the Markdown source file. It understands only a path relative to the
doxygen program working directory (which is the top-dir of the project).
GitLab on the other hand understands both document relative paths and
top-dir relative paths. As such, all internal document links in the
Markdown files are relative to the top-dir.

Unfortunately, GitHub only understands relative paths, but not top-dir
relative paths. This breaks document browsing from `README.md` and
`ldms/README.md` on GitHub.

This patch address the problem as follows:
  - Change all internal hyperlink references to be relative to its
    current file.
  - Apply a Doxygen input filter script to expand the relative paths to
    top-dir relative paths.