Add a `root` condition in ldms access check
Add a `root can do anything` access check condition in
`ldms_access_check()`. We need this patch for using `listen` and `auth`
configuration commands because the current `owner` uid/gid relied on the
ldms auth plugin of the default listening endpoint. But, when
introducing the `listen` command, the actual creation of endpoints will
be after `prdcr` config object creation. Hence, the owner will have to
be the euid/egid of the process. This will become an issue for `ovis`
authentication method as it resolves all authenticated connection as
root (0/0), which may or may not be the same as process euid/egid.