Fix ldmsd default auth processing
The default authentication object was set up after the configuration
file has done processing, but `prdcr_add` in the config file that did
not specify `auth` needed the default auth object to create the `prdcr`.
Hence, the default auth object (and its attributes) should be set up
during config file processing, when we knew that `set CLI` options are
done. If no config file is given, the default auth is set up in the
`ldmsd_init()` as usual.