Fix no reference to LDMSDPORT
LDMSDPORT is defined in `build/ldms/config.h`, and the file did not get
installed. Instead, the alternate version of the file
`build/ldms/src/ovis-ldms-config.h` is installed to avoid header file
collision. Hence, LDMSDPORT macro won't be resolved when compiling a
plugin outside of the ovis tree.

This patch change the use of LDMSDPORT to OVIS_LDMS_LDMSDPORT varient
that is defined in the installed header file ovis-ldms-config.h.