zap must not be freed by ldms (potential use-after-free)
In `__ldms_xprt_zap_new()`, the zap driver is obtained and is cached
by `__ldms_zap_get(); zap_get();` call chain. But, in the error path of
`__ldms_xprt_zap_new()`, the cached `zap` driver is freed, leading to
use-after-free for the next call of the same transport type. This patch
modify the error path to not free the cached zap driver.