Implement an end-to-end protocol for deleting a metric set
This change implements a message exchange and reference
counting to ensure that sets and remote buffer descriptors
are not freed until all references have been dropped. Because
some references (e.g. a RDMA MR) are remote, a message
exchange is required to inform the remote peer that the set
has been deleted, and a response to the sender that the set
is no longer in use at the remote peer.