configurable ib mad network sampler plugin ibnet
A number of supporting utilities are included:
* ldms_ibnet_schema_name provides the schema name that will be used for a given set of plugin configuration parameters. This is handy for configuring storage policies.

* ldms-ibnet-sampler-gen provides configuration files that spread the load of sampling
the ports of the input fat tree fabric description over the nodes listed in the input samplers file. It also generates optionally a lid/hostname mapping file useful for labeling data during analysis. It also generates optionally visualizations of the network hardware topology (not the routing tables) using graphviz.

* ldms-gen-lidfile.sh dumps a lid/guid/port mapping usable for small network systems with only one sampler.