Implement git-version-gen for ldms versioning
Implement git-version-gen for ldms versioning. The git-version-gen
script comes from the autoconf git repository at commit 8a09003
in build-aux/git-gersion-gen.

The LDMS project just needs to begin using annotated tags for its
version tags to work with this change. Right now the version will
show as "UNKNOWN" until an annotated tag is created at a previous
version.

The project will also need to decide on a consistent version prefix
Right now this patch assumes that the version prefix is "OVIS-".

For testing purposes, the author made an annotated tag like so:

 $ git tag -a OVIS-4.3.3.0test OVIS-4.3.3 -m "Test tag"

With that tag in place, we test that things are working by
running:

 $ sh autogen.sh
 [...]
 $ ./configure --version
 ovis-ldms configure 4.3.3.0test.79-6422
 generated by GNU Autoconf 2.69

We can see that the current commit "6422" is 79 commits ahead of the
4.3.3.0test tag.

Further we can test that the dist tarball is created successfully:

 $ ./configure
 [...]
 $ make dist
 $ ls ovis-ldms-*.tar.gz
 ovis-ldms-4.3.3.0test.79-6422.tar.gz

And if we unpack that tarball somewhere and verify the version within,
we see:

 /tmp/ovis-ldms-4.3.3.0test.79-6422$ ./configure --version
 ovis-ldms configure 4.3.3.0test.79-6422
 generated by GNU Autoconf 2.69

The patch author recommends that the Ovis ldms project force push
a replacement of the OVIS-4.3.3 tag with one that has the same name,
but is an annotated tag.

Fixes #95