Introduce automated testing of "make dist"
It is not uncommon for changes to the ovis build system to break the
"make dist" target. We introduce a GitHub Actions workflow to automate
the testing of the build system up through "make dist" so we will
be alerted of breakage on PRs for OVIS-4 and on the OVIS-4 branch.

The master branch has not yet had all of the build system cleanup that
OVIS-4 has, but this testing can easily be expaned to master once that
has happened.