Introduce create-release.yaml GitHub workflow
This introduces a new "Create Release" GitHub workflow, which automates
the process of creating a release on GitHub any time a new tag is pushed
to the repo.

The workflow triggers any time a new tag is pushed to the repo that
matches the pattern 'OVIS-4*'.

First the workflow builds the cannonical tarball.

Next the workflow generates a new release on the
https://github.com/ovis-hpc/ovis/releases page. The release notes
are taken from the body of the annotated tag.

Finally, the cannonical tarball is uploaded to the Assets of the
newly created release.