zap_sock - write before posting to sq
This patch modifies zap_sock internal sending mechanism to directly
write to the socket and post the the remainder to the send queue (sq).
Previously, all send data were posted to sq and could cause a send delay
as sq is handled by another zap_sock thread.

This patch passed `zap_test` and all of the test cases in `ldms-test`
repository.