Refactor the test_sampler plugin
Without the patch, the test_sampler plugin requests multiple 'config'
lines to define a schema and to add a set. This design is incompatible
with the new JSON configuration protocol. Moreover, the test_sampler
plugin was extended many times, and its code has become complicated.

The change simplifies the plugin code and makes it compatible with the
JSON protocol.