Prevent removing an RBD from the xprt rbd tree twice
An RBD could be in both a set RBD list and a transport RBD tree. This
results in a possible race between set deletion and disconnect paths.
__ldms_xprt_resource_free() unlocks the transport lock if an RBD
attaches to a set and relocks the transport lock after removing the RBD
from the set RBD list. __ldms_xprt_resource_free() may lose a set delete
race and then the RBD has been removed from the transport RBD tree in
the set deletion path. Without checking whether rbd->xprt is NULL or
not, we risk to removing the RBD from the set RBT tree twice and also
putting an extra transport reference.