Fix a crash when a config file contains a line longer than 8172.
When a line in a configuraton file is longer than the default maximum
message size of the config file cfg_xprt which is 8172 characters.
ldmsd_config.c:__process_config_file() would pass a record of the whole
request to ldmsd_config.c:__req_filter_failover() which assumes that it
always receives a whole request.

Instead of changing __req_filter_failover() , __process_config_file()
will always pass a whole request to __req_filter_failover(). This is to
avoid changing the failover logic and make a minimal change.