Fix ovis_json CFLAGS and compilation warnings
`CFLAGS := $(filter-out -Werror, ${CFLAGS})` statement in
`ovis_json/Makefile.am` was actually a Makefile statement. Automake took
that literally into Makefile that was generated by configure script.
Since `CFLAGS` was usually not defined in the environment variable nor
given when `make` was called, `CFLAGS` became an empty string.

Without the filter-out statement, the value of the CFLAGS in the
corresponding Makefile generated by configure is the value of CFLAGS
given in configure command.

This patch removed the filter-out statement that made CFLAGS an empty
string and addressed the build errors caused by `-Wall -Werror`.