Add ldms_set_{uid|gid|perm}_set() functions
Create ldms_set{uid|guid|perm}_set() functions, which are the inverse of
the existing ldms_set{uid|guid|perm}_get() functions.

The sampler plugins accept uid, gid and perm options. The user may supply
one or more of those options. However the existing API only allows the
plugin to supply all three options at once. This resulted in many plugin
authors implementing their own uid/gid/perm defaults for any of the
other options that were not supplied by the user. This results ldms metric
default auth values being hard coded in many different places in the code.

This is the first step in making it easy for sampler plugins to only set
the values for which they received configuration options and leave the
default to be set once in ldmsd proper.