Improve autoconf cython test
This version packs the cython test into a function.

The major change is that this version looks for variations on the
cython binary name that contain the python version string that
we detected. That way we don't need to hard code various python
version suffixes for cython. Also, we avoid the issue that there
may potentially be multiple different cython's installed for
multiple versions of Python at the same time (e.g. cython3.6 and
cython3.8 at the same time). This patch makes sure that we don't
pick the cython for the wrong version of python.

Also we run the cython version check in the feature check for
each cython that we find, not just the first one. That is minor
improvement as well.

Made suggested changes in issue #402.