Make the python & cython autoconf checks optional by default
Make the python & cython autoconf checks optional by default.
Also allow python & cython to be required with --enable-python,
and disabled with --disable-python.

It turns out that the AX_PYTHON_DEVEL macro is dependant on the
AM_PATH_PYTHON macro. While the AX_PYTHON_DEVEL macro does not
have action-if-found/-not-found options, the AM_PATH_PYTHON macro
does. While it this still allows that AM_PATH_PYTHON can succeed
and AX_PYTHON_DEVEL can fail, in most instances this won't be
the case.

The CYTHON_CHECK macro is upgraded to have action-if-found and
action-if-not-found options. Also the macro is moved into its
own file in the m4 directory.

Since the cython check is now conditional, we remove the special
cython handling from the make dist and create release workflows.