Verify the given interval and offset values of an updater are valid
The checks are
 * the given interval value is not an empty string,
 * the given interval value is a number,
 * the given interval value is greater than 0,
 * the given offset value is not an empty string,
 * the given offset value is a number,
 * the absolute value of the offset is not larger than the half of the
   interval value.

The checks in `updtr_add_handler` is to send the configuration error
messages to the log file or `ldmsd_controller`. The checks in
`ldmsd_updtr_new_with_auth` is to make sure that the given attribute
values are valid, especially when the caller is not `updtr_add_handler`.

Fixes #423