Fix double-free when LDMS lookup yields multiple sets
In some lookup mode, e.g. LOOKUP_RE and LOOKU_BY_SCHEMA, there can be
multiple sets resulting from the lookup operation. The lookup requesting
side shall expect multiple rendezvous messages from this one lookup
request. However, on the requesting side, the lookup request context was
prematurely freed after the first rendezvous message was received, and
the same ctxt was double-freed in the second rendezvous, and so-on.

The `lookup.more` flag in the rendezvous message tells the requesting
side if this is the last rendezvous messgae. It shall be checked, and
the lookup context shall not be freed if more rendezvous messages are
expected.