LDMS Cython with proper exceptions
LDMS Cython was using `RuntimeError` to raise an error from within the
module to the Python application. Even with string describing the error,
it lacks useful attribute such as `errno` from the underlying C code.

In this patch, `RuntimeError` usages were replaced with
`ConnectionError` if the error concerned a transport. For other errors
not directly concerning a transport but stemmed from C function calls,
`OSError` (the parent of all exceptions that have `errno` attribute) is
used with the corresponding `errno`. For all other errors not concerning
C function calls, the more appropriate exception classes from the Python
"builtins" were selected.