Introduce a simple first build test on CentOS7
Create a simple initial build test for LDMS. We can expand on this over time,
but this servers as a place to start. Use the Docker Hub container
morrone/centos7-lmds-build:latest, which is just the centos:7 base image
with a few of LDMS's build requirements add. The container  does include
the Cython package from EPEL.