Simple top level .gitignore fixes
Change the "build*" ignore rule to "/build*" so it applies at the top level,
not the entire tree. This will allow me to have a workflow named
"build-test-centos7.yaml", for instance.

Change the "configure" to be top-level as well, since we now should have
only one configure script in the tree.