Address `-Wall -Werror` error (format-truncation)
`-Wall` in GCC 9.3.0 enables `-Wformat` which also enables
`-Wformat-truncation`. The `-Wformat-truncation` warns about possible
formatted output truncationl. For example,
    char buff[64];
    char s[64];
    fgets(s, 63, stdin);
    snprintf(buff, sizeof(buff), "s is %s", s);
the length of `s` could be 63, the format "s is %s" could be
larger than 64 which results in a warning that turns in to an error by
`-Werror`.

The two places in the code tree that have format-truncation warnings
have the same kind of usage above, i.e. string printing into another
string with the same size. The printing strings (equivalent to `s` in
example above) were too big for the content they hold. Downsizing
them appropriately cleared the format-truncation warnings.