Fix zap_ugni GNI_SubscribeErrors error in non-privileged mode
Calling GNI_SubscribeErrors() with NULL nic_handle is restricted for
privileged users. For a non-privileged user, the function call results
in GNI_RC_INVALID_PARAM. Since GNI_SubscribeErrors() is called after
zap_ugni attached to a domain and got a `_dom.nic` handle, the handle is
is used instead of NULL so that non-privileged users can also use
zap_ugni.