Fix ldmsctl crashes when it receives a multiple-record message
Without the fix, ldmsctl reused a single buffer to receive message
records. It assumed that the buffer would be handled before the next
record arrives. The assumption is not always true if a response message
consists of multiple records. The symptom is as if ldmsctl does not
receive some records, resulting in ldmsctl crashes.