Add test for annotated tag to create-release workflow, other cleanup
Add a test that uses git for-each-ref to test if the tag looks like
a annotated tag. If it does not, we print a clear error message,
"does not look like an annotaded tag".

Also, github was warning that set-env is deprecated, so we update
to using the "$GITHUB_ENV" file method of setting environment
variables for later steps.

Finally, update to using the ubuntu-20.04 VM from ubuntu-18.04.