Fix zap_fabric:get_name() so that it can receive sa_len of 0.
The sa_len parameters of the zap_get_name() API is an output parameter.
The given sa_len value could be 0. Without the change, zap_fabric() would
return an error if the value is 0. This change makes the behavior of
zap_fabric's get_name() compatible with those in the other zap
transports.