Fix zap endpoint leak
Zap endpoint `x->zap_ep` in the LDMS transport were set to `NULL` in
`__ldsm_xprt_resource_free()` without calling `zap_free()`. As a
result, the zap endpoint is leaked. The bug is verified in sock, rdma
and fabric-over-verbs transports by running a sampler ldmsd and check
the number of file descriptors before and after `ldms_ls` to the
daemon.

This patch freed the zap endpoint before setting it to `NULL`. The fix
is verified with the file descriptors check described above on sock,
rdma and fabric-over-verbs transports.