Fix endpoint leak when zap_connect failed synchronously
When `zap_connect()` failed synchronously, zap endpoint were left alone
and the ldms xprt reference that was taken when setting zap ucontext was
not put back. Hence, the LDMS xprt reference would never reached 0 and
the associated resources would not be freed. This patch releases the
zap endpoint and put down the associated LDMS xprt reference.