Fix zap endpoint leak in connect error event
When ldms xprt received zap connect error event, it set `xprt->zap_ep`
to `NULL` without releasing the zap endoint. This leads to zap endpoint
leak (and file descriptor leak). The bug can be verified by running an
aggregator with a prdcr that tries to connect to a sampler that is not
running. The prdcr will keep reconnecting, and hence growing
/proc/PID/fd. This patch simply put in the forgotten `zap_free()` to
release the zap endpoint.