Improve the dir_update path
* Ignore the json_data_len value in ldms_dir_reply structure
There is a bug in the dir path in OVIS V4.3.3 and the bug has been fixed
in V4.3.4. The fix makes the two versions incompatible. The change makes
the two versions compatible.

 * Allocate the dir_update buffer of the needed size
Before the change, the buffer was allocated as large as the transport's
maximum message. The maximum message size of uGNI and Sock is 1MB which
is very likely to it is much larger than what dir_update message size
would be.

The changes allocates the memory of dir_update message exactly what it
needs.