Fix `--with-slurm` error checking (typo)
When `--with-slurm=/some/path` is specified and `slurm/spank.h` header
is not found, the configure script should report an error and exit.
However, due to a typological error, the condition that tests for
"no slurm" case is always false and no error is reported. `have_slurm`
is still a `no`. So, the `slurm_notifier` plugin is silently disabled.
Removing the typological error (extra '=') solved the problem. The
configure script now error-exit if `--with-slurm` is given and the
test for `slurm/spank.h` failed.