Fix zap ep leak in ldms xprt
The zap references taken when LDMS xprt sends a lookup message or a dir
message are put back when the corresponding reply messages have
received. However, the reply messages may never arrive because of the
disconnection and zap endpoints are leaked. These zap references before
send are not necessary and shall be removed.