Fix zap.c so that LDMS processes can run as daemons on Ubuntu20.04
When an LDMSD starts on Ubuntu20.04 as a daemon, the process stuck after
the demonization process starts. It turns out that it is a deadlock when
glibc is handling the child hook registered in zap.c by calling
pthead_atfork(). In zap.c, cs_init() calls pthread_atfork() and
registers cs_init() itself as the child hook handler.