Prevent SIGPIPE in zap_sock
zap_sock used `write()` to write messages to the socket. It is possible
to write to a disconnected socket, and when it happened SIGPIPE was
sent to the calling process and the process terminated (default SIGPIPE
action). `SO_NOSIGPIPE` socket option is also not available in Linux
(see `linux/include/uapi/asm-generic/socket.h`).

Instead of `write()`, this patch uses `send()` with `MSG_NOSIGNAL` to
disable the SIGPIPE on disconnected socket. This patch also bundles
various socket options setting into one function for the ease of
maintenance.