remove buggy ldmsd_msg_logger function
the ldmsd_msg_logger function:
is not thread safe
is passed to all the plugins that take a log pointer
contains a buffer overrun
is redundant with ldmsd_log.

This patch removes the function and redirects the only caller to ldmsd_log.
It leaves a #define for any client code that might have used it.