Fix a possible use-after-freed in the cleanup path
The cleanup logic iterates through the storage policy tree to close all
stores. It also puts back a storage policy reference, which is
erroneous. It is unnecessary to put the reference back as we only want
to close the stores and not free the storage policies. Moreover, this
could cause a use-after-free if all producer sets associated with a
storage policy have been deleted. The storage policy will be freed when
the cleanup logic puts back the reference, and then the cleanup logic
reaccesses the storage policy to get the next storage policy.