Fix rbd use after free
The releasing of rbd in __ldms_xprt_resource_free() could race with
ldms_xprt_set_delete(). In such case, __ldms_xprt_resource_free()
released the xprt lock, and then ldms_xprt_set_delete() took the lock
and call __ldms_rbd_xprt_release() which removed the rbd from the tree
and drop a reference. That reference could be the last reference to
the rbd, and hence once ldms_xprt_set_delete() released the lock and
__ldms_xprt_resource_free() took the lock again, the rbd would already
be freed -- causing a use-after-free issue.

This patch solved the issue by taking the rbd reference while the rbd
is still in the tree and xprt lock is still hold, and drop the
reference after we're done processing it.