Import m4 files from gnulib instead of gettext
gettext is not the true source of config.rpath,, lib-ld.m4,
lib-link.m4, and lib-prefix.m4. gnulib is the true source of those
files. This commit updates ovis to get the files from the true origin
rather than getting them indirectly from another project.

This approach both gets us more up-to-date files right now and also
give us a cleaner documented approach for updating an maintaining
those file in the future.

The gnulib-tool command from the gnulib project is used to import
the current latest m4 files in support of AC_LIB_HAVE_LINKFLAGS.
This also has the effect of including the AC_LIB_RPATH macro.

This was run when the gnulib project was at commit

  db493d1ce43fc39dd925699d840c80b3400f4313

in its repository.

The following command was employed:

  gnulib-tool --no-vc-files --source-base=gnulib --import havelib