Throw an error if auth enabled but openssl not found
The ovis_auth library has a hard dependency on openssl.  Change the configure
script to throw an error if ovis-auth is enabled, but openssl is not found.

Remove the HAVE_SSL variable, which was unused and would currently serve
little purpose.

Update ax_check_openssl.m4 to the latext from autoconf-archive to resolve
issue when using CFLAGS="-Wall -Werror".

Fixes #80