Search for the push RBD in the push path and fix the push RBD references
Upon receiving a push register request, LDMS calls ldms_lookup_rbd() to
search for an RBD corresponding to an LDMS transport and a set. If the
client has registered for push updates before, at least 2 RBDs
correspond to the endpoint and the set are in x->rbt_rbd. The RBD
returned by ldms_lookup_rbd() could be any of the RBDs. If a client
registers for push multiple times, multiple RBDs could be created
instead of updating an existing one. In this case, the server will send
multiple push updates to the client when it calls ldms_xprt_push(). When
handling a cancel_push request, the server may not cancel the push
registration at all if ldms_lookup_rbd() returns an RBD that is not a
push RBD.

Moreover, LDMS takes a reference every time it receives a push register
request. It is unnecessary. Upon receiving a cancel push request, LDMS
calls __ldms_free_rbd(), which in turn puts back the “rendezvous_push”
references twice. It does not match the number of “rendezvous_push”
references.