Add enumeration compatibility checking
This commit add a github action checking that the future pull requests
preserve the values of enumerations in:
 - `.github/enum-check/ovis-4.3.3-enum.list` compared to OVIS-4.3.3,
 - `.github/enum-check/ovis-4.3.4-enum.list` compared to OVIS-4.3.4,
 - `.github/enum-check/ovis-4.3.5-enum.list` compared to OVIS-4.3.5,
 - `.github/enum-check/ovis-4.3.6-enum.list` compared to OVIS-4.3.6,

Please note that `LDMSD_LISTEN_REQ` is dropped from ovis-4.3.4-enum.list
and ovis-4.3.5-enum.list because OVIS-4.3.6 changed the value. However,
`LDMSD_LISTEN_REQ` is included in ovis-4.3.6-enum.list to make sure that
the future pull requests do not change it by accident.