Unify the code that chunks large messages into multiple records
The logic to chunk messages that are larger than the maximum message
length the transport supports was implemented in multiple places. It is
hard to maintain the code.

The change introduces the ldmsd_msg_buf_send() API in the ldmsd_request
library. The API appends data to a caller’s buffer and sends the data
when the buffer is full or the caller has passed the whole message,
i.e., callers pass the LDMSD_REQ_EOM_F flag. ldmsd_request and
ldmsd_stream use the API instead of implementing their own functions.