Create ldms_set_default_authz() function
Create a new ldms_set_default_authz() function which can be used to
set and/or get the default authorization values for metric sets.  This
is implemented as a single function rather than separate set/get
functions because setting the three values needs to be atomic, and
one might want to know what the resulting defaults are after setting
only one or two of the values. That retrieval (to be completely
pedantic) needs to be atomic as well. So this function can return
the existing defaults in the fields which are not currently being
used to _set_ values. And since it can do that, then it can also
be used in readonly mode to get all of the values.

Move the initialization of the default metric sets authorization values
to ldms_init().

Update some related comments for ldms_set_new() and ldms_set_new_with_auth().